+++ api/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "testone",
+  "name": "api",
   "version": "1.0.0",
   "description": "",
   "main": "index.js",
@@ -11,6 +11,14 @@
   "dependencies": {
     "cors": "^2.8.5",
     "express": "^4.18.1",
-    "mongoose": "^6.3.4"
+    "mssql": "^8.1.2"
+  },
+  "devDependencies": {
+    "@types/cors": "^2.8.12",
+    "@types/express": "^4.17.13",
+    "@types/node": "^17.0.37",
+    "nodemon": "^2.0.16",
+    "ts-node": "^10.8.0",
+    "typescript": "^4.7.2"
   }
 }
