+++ api/src/conf/db.ts
@@ -1,0 +1,18 @@
+//import mssql from 'mssql';
+/*
+const pool = new mssql.ConnectionPool({
+    user: process.env.DB_USER || "a",
+    password: process.env.DB_PASS || "a",
+    server: process.env.DB_SERVER || "a",
+    port: process.env.DB_PORT || "a",
+    database: process.env.DB_NAME || "a",
+    options: {
+        encrypt: false,
+        trustServiceCertificate: false
+    }
+}).connect;
+
+//const pool = new mssql.ConnectionPool(connectionString).connect;
+
+export default pool;
+*/
