+++ api/src/controllers/auth.ts
@@ -1,0 +1,1 @@
+//
