+++ api/src/models/auth.ts
@@ -1,0 +1,1 @@
+//
