+++ api/src/routes/auth.ts
@@ -1,0 +1,1 @@
+//
